+++ src/Graphics/RenderTarget.cpp
@@ -68,7 +68,7 @@ Texture2D* RenderTarget::GetTexture()
 
 void RenderTarget::Clear(Color clearColor)
 {
-	GLfloat color[] = {clearColor.R/255.f, clearColor.G/255.f, clearColor.B/255.f, clearColor.A/255.f};
+	GLfloat color[] = {GLfloat(clearColor.R)/255.f, GLfloat(clearColor.G)/255.f, GLfloat(clearColor.B)/255.f, GLfloat(clearColor.A)/255.f};
 	glClearColor(color[0], color[1], color[2], color[3]);
 	glClear(GL_COLOR_BUFFER_BIT);
 }
